Easier to remember password generator
Use the txt file as a reference list. 

Txt file contains names of animals (non-latin words) and 1000 of the most common English words. 

This program is used as a reference to make passwords that are easy to remember. I'm currently trying to add special symbols and random capitalization in the beginning of a word.